refactor(web): split catalog page routes into modules [OPE-299]
